requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
azure-ai-textanalytics==5.3.0
azure-cognitiveservices-speech==1.38.0
azure-ai-vision-imageanalysis==1.0.0
python-dotenv==1.0.1
streamlit==1.35.0
newsapi-python==0.2.7
